refactor: Move description to trait
